폴더 및 파일 구조 리펙토링 : refactor : 기능별 코드 분리를 위한 리펙토링 https://github.com/Chuseok22/ScreensaverAuditor/issues/12
